+++ src/Gruntz/Multi.cpp
@@ -621,8 +621,7 @@ i32 CMulti::PumpA() {
     g_killCueClock = g_lastNow;
     g_engineFrameDelta = 0x21;
     if (m_ambientInitDone == 0) {
-        if (static_cast<i64>(static_cast<u32>(g_frameTime)) - m_ambientTimer64.m_v
-            >= m_ambientInterval64.m_v) {
+        if (static_cast<i64>(g_frameTime) - m_ambientTimer64.m_v >= m_ambientInterval64.m_v) {
             char name[0x40];
             wsprintfA(name, "AMBIENT%d", GetAmbientId());
             if (g_gameReg->m_musicEnabled != 0) {
